requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
reportlab==4.0.9
httpx
python-dotenv==0.19.0
aiofiles
cffi
pymupdf
faiss-cpu
cashfree_pg
cashfree_verification
python-decouple
pandas 
gspread 
oauth2client